controller add exception try catch

+++ src/controller/api.ts
@@ -1,5 +1,3 @@
-
-
 import {
   Inject,
   Controller,
@@ -12,15 +10,15 @@ import {
   Put,
   Del,
   Validate,
- } from '@midwayjs/decorator';
-import { Context } from 'egg';
-import { IGetTodoItemResponse } from '../interface';
-import { TodoItemService } from '../service/todoItem';
-import { TNgaTodoItem } from '../entity/TNgaTodoItem';
+} from '@midwayjs/decorator';
+import {Context} from 'egg';
+import {IGetTodoItemResponse} from '../interface';
+import {TodoItemService} from '../service/todoItem';
+import {TNgaTodoItem} from '../entity/TNgaTodoItem';
 
 
 @Provide()
-@Controller('/api/todo', { middleware: [ 'reportMiddleware' ] })
+@Controller('/api/todo', {middleware: ['reportMiddleware']})
 export class APIController {
   @Inject()
   ctx: Context;
@@ -28,38 +26,69 @@ export class APIController {
   @Inject()
   todoItemService: TodoItemService;
 
+  @Inject()
+  logger;
+
   @Get('/get')
   async get(@Query() id: number): Promise<IGetTodoItemResponse> {
-    if (id == null)
-      throw new Error('id is null');
-    const todoItems = await this.todoItemService.get(id);
-    return { success: true, message: 'OK', data: todoItems };
+
+    if (id == null || typeof id !== "number") {
+      this.logger.warn('id is null')
+      return {success: false, message: 'id is null', data: null}
+    }
+    try {
+      const todoItems = await this.todoItemService.get(id);
+      return {success: true, message: 'OK', data: todoItems};
+    } catch (e) {
+      return {success: false, message: 'ERROR', data: e}
+    }
   }
 
   @Post('/save')
   @Validate()
   async save(@Body(ALL) todoItem: TNgaTodoItem): Promise<IGetTodoItemResponse> {
-    const result = await this.todoItemService.save(todoItem);
-    return { success: true, message: 'OK', data: result };
+    try {
+      const result = await this.todoItemService.save(todoItem);
+      return {success: true, message: 'OK', data: result};
+    } catch (e) {
+      return {success: false, message: 'ERROR', data: e}
+    }
   }
 
+
   @Put('/put')
+  @Validate()
   async put(@Body(ALL) todoItem: TNgaTodoItem): Promise<IGetTodoItemResponse> {
-    const result = await this.todoItemService.put(todoItem);
-    return { success: true, message: 'OK', data: result };
+    try {
+      const result = await this.todoItemService.put(todoItem);
+      return {success: true, message: 'OK', data: result};
+    } catch (e) {
+      return {success: false, message: 'ERROR', data: e}
+    }
   }
+
   @Get('/list')
   async list(): Promise<IGetTodoItemResponse> {
-    const todoItems = await this.todoItemService.list();
-    return { success: true, message: 'OK', data: todoItems };
+    try {
+      const todoItems = await this.todoItemService.list();
+      return {success: true, message: 'OK', data: todoItems};
+    } catch (e) {
+      return {success: false, message: 'ERROR', data: e}
+    }
   }
 
   @Del('/delete')
   async delete(@Query() id: number): Promise<IGetTodoItemResponse> {
-    if (id == null )
-      throw new Error('id is null');
-    const todoItems = await this.todoItemService.delete(id);
-    return { success: true, message: 'OK', data: todoItems };
+    if (id == null || typeof id !== "number") {
+      this.logger.warn('id is null')
+      return {success: false, message: 'id is null', data: null}
+    }
+    try {
+      const todoItems = await this.todoItemService.delete(id);
+      return {success: true, message: 'OK', data: todoItems};
+    } catch (e) {
+      return {success: false, message: 'ERROR', data: e}
+    }
   }
